gpio/requirements.txt
paho.mqtt==2.0.0
RPi.GPIO==0.7.1